renamed vec113 package
