+++ vec113/pom.xml
@@ -124,7 +124,7 @@
                                 <arg>-Xvisitor</arg>
                                 <arg>-Xvisitor-includeType</arg>
                                 <arg>-Xvisitor-noIdrefTraversal</arg>
-                                <arg>-Xvisitor-package:com.foursoft.vecmodel.vec113.visitor</arg>
+                                <arg>-Xvisitor-package:com.foursoft.harness.vec.v113.visitor</arg>
                             </args>
                             <generateDirectory>${project.build.directory}/generated-sources/xjc-vec113</generateDirectory>
                         </configuration>
