+++ vec113/src/main/java/com/foursoft/harness/vec/v113/VecReader.java
@@ -23,7 +23,7 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.vecmodel.vec113;
+package com.foursoft.harness.vec.v113;
 
 import com.foursoft.xml.io.read.XMLReader;
 import com.foursoft.xml.model.Identifiable;
